refactor: split and decouple code files
